+++ src/all/misskon/src/eu/kanade/tachiyomi/extension/all/misskon/Filters.kt
@@ -1,128 +1,129 @@
 package eu.kanade.tachiyomi.extension.all.misskon
 
-import eu.kanade.tachiyomi.extension.all.misskon.SourceCategorySelector.Companion.CategoryPresets.CHINESE
-import eu.kanade.tachiyomi.extension.all.misskon.SourceCategorySelector.Companion.CategoryPresets.KOREAN
-import eu.kanade.tachiyomi.extension.all.misskon.SourceCategorySelector.Companion.CategoryPresets.OTHER
-import eu.kanade.tachiyomi.extension.all.misskon.SourceCategorySelector.Companion.CategoryPresets.TOP
 import eu.kanade.tachiyomi.source.model.Filter
 
-class SourceCategory(private val name: String, val url: String) {
-    override fun toString() = this.name
+class TopDays(val name: String, val uri: String)
+
+class TopDaysFilter(displayName: String, private val days: Array<TopDays>) : Filter.Select<String>(displayName, days.map { it.name }.toTypedArray()) {
+    fun toUriPart() = days[state].uri
 }
 
-class SourceCategorySelector(
-    name: String,
-    categories: List<SourceCategory>,
-) : Filter.Select<SourceCategory>(name, categories.toTypedArray()) {
+fun getTopDaysList() = arrayOf(
+    TopDays("<Select>", ""),
+) + topDaysList()
 
-    val selectedCategory: SourceCategory?
-        get() = if (state > 0) values[state] else null
+fun topDaysList() = arrayOf(
+    TopDays("Top 3 days", "top3"),
+    TopDays("Top week", "top7"),
+    TopDays("Top month", "top30"),
+    TopDays("Top 2 months", "top60"),
+)
 
-    companion object {
-
-        private object CategoryPresets {
-
-            val TOP = listOf(
-                SourceCategory("Top 3 days", "/top3/"),
-                SourceCategory("Top 7 days", "/top7/"),
-                SourceCategory("Top 30 days", "/top30/"),
-                SourceCategory("Top 60 days", "/top60/"),
-            )
-
-            val CHINESE = listOf(
-                SourceCategory("Chinese:[MTCos] 喵糖映画", "/tag/mtcos/"),
-                SourceCategory("Chinese:BoLoli", "/tag/bololi/"),
-                SourceCategory("Chinese:CANDY", "/tag/candy/"),
-                SourceCategory("Chinese:FEILIN", "/tag/feilin/"),
-                SourceCategory("Chinese:FToow", "/tag/ftoow/"),
-                SourceCategory("Chinese:GIRLT", "/tag/girlt/"),
-                SourceCategory("Chinese:HuaYan", "/tag/huayan/"),
-                SourceCategory("Chinese:HuaYang", "/tag/huayang/"),
-                SourceCategory("Chinese:IMISS", "/tag/imiss/"),
-                SourceCategory("Chinese:ISHOW", "/tag/ishow/"),
-                SourceCategory("Chinese:JVID", "/tag/jvid/"),
-                SourceCategory("Chinese:KelaGirls", "/tag/kelagirls/"),
-                SourceCategory("Chinese:Kimoe", "/tag/kimoe/"),
-                SourceCategory("Chinese:LegBaby", "/tag/legbaby/"),
-                SourceCategory("Chinese:MF", "/tag/mf/"),
-                SourceCategory("Chinese:MFStar", "/tag/mfstar/"),
-                SourceCategory("Chinese:MiiTao", "/tag/miitao/"),
-                SourceCategory("Chinese:MintYe", "/tag/mintye/"),
-                SourceCategory("Chinese:MISSLEG", "/tag/missleg/"),
-                SourceCategory("Chinese:MiStar", "/tag/mistar/"),
-                SourceCategory("Chinese:MTMeng", "/tag/mtmeng/"),
-                SourceCategory("Chinese:MyGirl", "/tag/mygirl/"),
-                SourceCategory("Chinese:PartyCat", "/tag/partycat/"),
-                SourceCategory("Chinese:QingDouKe", "/tag/qingdouke/"),
-                SourceCategory("Chinese:RuiSG", "/tag/ruisg/"),
-                SourceCategory("Chinese:SLADY", "/tag/slady/"),
-                SourceCategory("Chinese:TASTE", "/tag/taste/"),
-                SourceCategory("Chinese:TGOD", "/tag/tgod/"),
-                SourceCategory("Chinese:TouTiao", "/tag/toutiao/"),
-                SourceCategory("Chinese:TuiGirl", "/tag/tuigirl/"),
-                SourceCategory("Chinese:Tukmo", "/tag/tukmo/"),
-                SourceCategory("Chinese:UGIRLS", "/tag/ugirls/"),
-                SourceCategory("Chinese:UGIRLS - Ai You Wu App", "/tag/ugirls-ai-you-wu-app/"),
-                SourceCategory("Chinese:UXING", "/tag/uxing/"),
-                SourceCategory("Chinese:WingS", "/tag/wings/"),
-                SourceCategory("Chinese:XiaoYu", "/tag/xiaoyu/"),
-                SourceCategory("Chinese:XingYan", "/tag/xingyan/"),
-                SourceCategory("Chinese:XIUREN", "/tag/xiuren/"),
-                SourceCategory("Chinese:XR Uncensored", "/tag/xr-uncensored/"),
-                SourceCategory("Chinese:YouMei", "/tag/youmei/"),
-                SourceCategory("Chinese:YouMi", "/tag/youmi/"),
-                SourceCategory("Chinese:YouMi尤蜜", "/tag/youmiapp/"),
-                SourceCategory("Chinese:YouWu", "/tag/youwu/"),
-            )
-
-            val KOREAN = listOf(
-                SourceCategory("Korean:AG", "/tag/ag/"),
-                SourceCategory("Korean:Bimilstory", "/tag/bimilstory/"),
-                SourceCategory("Korean:BLUECAKE", "/tag/bluecake/"),
-                SourceCategory("Korean:CreamSoda", "/tag/creamsoda/"),
-                SourceCategory("Korean:DJAWA", "/tag/djawa/"),
-                SourceCategory("Korean:Espacia Korea", "/tag/espacia-korea/"),
-                SourceCategory("Korean:Fantasy Factory", "/tag/fantasy-factory/"),
-                SourceCategory("Korean:Fantasy Story", "/tag/fantasy-story/"),
-                SourceCategory("Korean:Glamarchive", "/tag/glamarchive/"),
-                SourceCategory("Korean:HIGH FANTASY", "/tag/high-fantasy/"),
-                SourceCategory("Korean:KIMLEMON", "/tag/kimlemon/"),
-                SourceCategory("Korean:KIREI", "/tag/kirei/"),
-                SourceCategory("Korean:KiSiA", "/tag/kisia/"),
-                SourceCategory("Korean:Korean Realgraphic", "/tag/korean-realgraphic/"),
-                SourceCategory("Korean:Lilynah", "/tag/lilynah/"),
-                SourceCategory("Korean:Lookas", "/tag/lookas/"),
-                SourceCategory("Korean:Loozy", "/tag/loozy/"),
-                SourceCategory("Korean:Moon Night Snap", "/tag/moon-night-snap/"),
-                SourceCategory("Korean:Paranhosu", "/tag/paranhosu/"),
-                SourceCategory("Korean:PhotoChips", "/tag/photochips/"),
-                SourceCategory("Korean:Pure Media", "/tag/pure-media/"),
-                SourceCategory("Korean:PUSSYLET", "/tag/pussylet/"),
-                SourceCategory("Korean:SAINT Photolife", "/tag/saint-photolife/"),
-                SourceCategory("Korean:SWEETBOX", "/tag/sweetbox/"),
-                SourceCategory("Korean:UHHUNG MAGAZINE", "/tag/uhhung-magazine/"),
-                SourceCategory("Korean:UMIZINE", "/tag/umizine/"),
-                SourceCategory("Korean:WXY ENT", "/tag/wxy-ent/"),
-                SourceCategory("Korean:Yo-U", "/tag/yo-u/"),
-            )
-
-            val OTHER = listOf(
-                SourceCategory("Other:AI Generated", "/tag/ai-generated/"),
-                SourceCategory("Other:Cosplay", "/tag/cosplay/"),
-                SourceCategory("Other:JP", "/tag/jp/"),
-                SourceCategory("Other:JVID", "/tag/jvid/"),
-                SourceCategory("Other:Patreon", "/tag/patreon/"),
-            )
-        }
-
-        fun create(): SourceCategorySelector {
-            val options = mutableListOf(SourceCategory("<Select>", "")).apply {
-                addAll(TOP)
-                addAll(CHINESE)
-                addAll(KOREAN)
-                addAll(OTHER)
-            }
-            return SourceCategorySelector("Category", options)
-        }
-    }
+class TagsFilter(displayName: String, private val tags: List<Pair<String, String>>) : Filter.Select<String>(displayName, tags.map { it.first }.toTypedArray()) {
+    fun toUriPart() = tags[state].second
 }
+
+// console.log([...document.querySelectorAll(".tag-counterz a")].map((el) => `Tag("${el.innerText.trim()}", "${decodeURIComponent(el.getAttribute('href'))}"),`).join('\n'))
+val DefaultTagList = setOf(
+    Pair("<Select>", ""),
+    Pair("--<Chinese>--", ""),
+    Pair("[MTCos] 喵糖映画", "mtcos"),
+    Pair("BoLoli", "bololi"),
+    Pair("CANDY", "candy"),
+    Pair("FEILIN", "feilin"),
+    Pair("FToow", "ftoow"),
+    Pair("GIRLT", "girlt"),
+    Pair("HuaYan", "huayan"),
+    Pair("HuaYang", "huayang"),
+    Pair("IESS异思趣向", "iess异思趣向"),
+    Pair("IMISS", "imiss"),
+    Pair("ISHOW", "ishow"),
+    Pair("ISS系列", "iss系列"),
+    Pair("JVID", "jvid"),
+    Pair("KelaGirls", "kelagirls"),
+    Pair("Kimoe", "kimoe"),
+    Pair("LegBaby", "legbaby"),
+    Pair("Limerence原创", "limerence原创"),
+    Pair("MF", "mf"),
+    Pair("MFStar", "mfstar"),
+    Pair("MiiTao", "miitao"),
+    Pair("MintYe", "mintye"),
+    Pair("MISSLEG", "missleg"),
+    Pair("MiStar", "mistar"),
+    Pair("MTMeng", "mtmeng"),
+    Pair("MyGirl", "mygirl"),
+    Pair("MZSOCK爱美足", "mzsock爱美足"),
+    Pair("PartyCat", "partycat"),
+    Pair("QingDouKe", "qingdouke"),
+    Pair("ROSI写真", "rosi写真"),
+    Pair("RuiSG", "ruisg"),
+    Pair("SLADY", "slady"),
+    Pair("TASTE", "taste"),
+    Pair("TGOD", "tgod"),
+    Pair("TouTiao", "toutiao"),
+    Pair("TuiGirl", "tuigirl"),
+    Pair("Tukmo", "tukmo"),
+    Pair("UGIRLS", "ugirls"),
+    Pair("UGIRLS - Ai You Wu App", "ugirls-ai-you-wu-app"),
+    Pair("Ugirls爱尤物", "ugirls-app"),
+    Pair("UXING", "uxing"),
+    Pair("WingS", "wings"),
+    Pair("XiaoYu", "xiaoyu"),
+    Pair("XingYan", "xingyan"),
+    Pair("XIUREN", "xiuren"),
+    Pair("XR Uncensored", "xr-uncensored"),
+    Pair("YouMei", "youmei"),
+    Pair("YouMi", "youmi"),
+    Pair("YouMi尤蜜", "youmiapp"),
+    Pair("YouWu", "youwu"),
+    Pair("内购无水印", "内购无水印"),
+    Pair("她们印象", "她们印象"),
+    Pair("精选街拍作品", "精选街拍作品"),
+
+    Pair("--<Korean>--", ""),
+    Pair("AG", "ag"),
+    Pair("Bimilstory", "bimilstory"),
+    Pair("BLUECAKE", "bluecake"),
+    Pair("CreamSoda", "creamsoda"),
+    Pair("DJAWA", "djawa"),
+    Pair("Espacia Korea", "espacia-korea"),
+    Pair("Fantasy Factory", "fantasy-factory"),
+    Pair("Fantasy Story", "fantasy-story"),
+    Pair("Glamarchive", "glamarchive"),
+    Pair("Haivia", "haivia"),
+    Pair("Herovia", "herovia"),
+    Pair("HIGH FANTASY", "high-fantasy"),
+    Pair("KIMLEMON", "kimlemon"),
+    Pair("KIREI", "kirei"),
+    Pair("KiSiA", "kisia"),
+    Pair("Korean Realgraphic", "korean-realgraphic"),
+    Pair("LE", "le"),
+    Pair("Lilynah", "lilynah"),
+    Pair("Lookas", "lookas"),
+    Pair("Loozy", "loozy"),
+    Pair("MakeModel", "makemodel"),
+    Pair("Moon Night Snap", "moon-night-snap"),
+    Pair("Paranhosu", "paranhosu"),
+    Pair("PhotoChips", "photochips"),
+    Pair("Pure Media", "pure-media"),
+    Pair("PUSSYLET", "pussylet"),
+    Pair("SAINT Photolife", "saint-photolife"),
+    Pair("Sera", "sera"),
+    Pair("SWEETBOX", "sweetbox"),
+    Pair("UHHUNG MAGAZINE", "uhhung-magazine"),
+    Pair("UMIZINE", "umizine"),
+    Pair("WXY ENT", "wxy-ent"),
+    Pair("Yo-U", "yo-u"),
+
+    Pair("--<Other>--", ""),
+    Pair("AI Generated", "ai-generated"),
+    Pair("AIGirl", "aigirl"),
+    Pair("Cosplay", "cosplay"),
+    Pair("JP", "jp"),
+    Pair("JVID", "jvid"),
+    Pair("OtherXXX", "otherxxx"),
+    Pair("Patreon", "patreon"),
+    Pair("Private Photoshoot", "private-photoshoot"),
+
+    Pair("--<More>--", ""),
+)
