--- src/all/misskon/src/eu/kanade/tachiyomi/extension/all/misskon/MissKonTags.kt
@@ -1,129 +1,0 @@
-package eu.kanade.tachiyomi.extension.all.misskon
-
-import eu.kanade.tachiyomi.source.model.Filter
-
-class TopDays(val name: String, val uri: String)
-
-class TopDaysFilter(displayName: String, private val days: Array<TopDays>) : Filter.Select<String>(displayName, days.map { it.name }.toTypedArray()) {
-    fun toUriPart() = days[state].uri
-}
-
-fun getTopDaysList() = arrayOf(
-    TopDays("<Select>", ""),
-) + topDaysList()
-
-fun topDaysList() = arrayOf(
-    TopDays("Top 3 days", "top3"),
-    TopDays("Top week", "top7"),
-    TopDays("Top month", "top30"),
-    TopDays("Top 2 months", "top60"),
-)
-
-class TagsFilter(displayName: String, private val tags: List<Pair<String, String>>) : Filter.Select<String>(displayName, tags.map { it.first }.toTypedArray()) {
-    fun toUriPart() = tags[state].second
-}
-
-// console.log([...document.querySelectorAll(".tag-counterz a")].map((el) => `Tag("${el.innerText.trim()}", "${decodeURIComponent(el.getAttribute('href'))}"),`).join('\n'))
-val DefaultTagList = setOf(
-    Pair("<Select>", ""),
-    Pair("--<Chinese>--", ""),
-    Pair("[MTCos] 喵糖映画", "mtcos"),
-    Pair("BoLoli", "bololi"),
-    Pair("CANDY", "candy"),
-    Pair("FEILIN", "feilin"),
-    Pair("FToow", "ftoow"),
-    Pair("GIRLT", "girlt"),
-    Pair("HuaYan", "huayan"),
-    Pair("HuaYang", "huayang"),
-    Pair("IESS异思趣向", "iess异思趣向"),
-    Pair("IMISS", "imiss"),
-    Pair("ISHOW", "ishow"),
-    Pair("ISS系列", "iss系列"),
-    Pair("JVID", "jvid"),
-    Pair("KelaGirls", "kelagirls"),
-    Pair("Kimoe", "kimoe"),
-    Pair("LegBaby", "legbaby"),
-    Pair("Limerence原创", "limerence原创"),
-    Pair("MF", "mf"),
-    Pair("MFStar", "mfstar"),
-    Pair("MiiTao", "miitao"),
-    Pair("MintYe", "mintye"),
-    Pair("MISSLEG", "missleg"),
-    Pair("MiStar", "mistar"),
-    Pair("MTMeng", "mtmeng"),
-    Pair("MyGirl", "mygirl"),
-    Pair("MZSOCK爱美足", "mzsock爱美足"),
-    Pair("PartyCat", "partycat"),
-    Pair("QingDouKe", "qingdouke"),
-    Pair("ROSI写真", "rosi写真"),
-    Pair("RuiSG", "ruisg"),
-    Pair("SLADY", "slady"),
-    Pair("TASTE", "taste"),
-    Pair("TGOD", "tgod"),
-    Pair("TouTiao", "toutiao"),
-    Pair("TuiGirl", "tuigirl"),
-    Pair("Tukmo", "tukmo"),
-    Pair("UGIRLS", "ugirls"),
-    Pair("UGIRLS - Ai You Wu App", "ugirls-ai-you-wu-app"),
-    Pair("Ugirls爱尤物", "ugirls-app"),
-    Pair("UXING", "uxing"),
-    Pair("WingS", "wings"),
-    Pair("XiaoYu", "xiaoyu"),
-    Pair("XingYan", "xingyan"),
-    Pair("XIUREN", "xiuren"),
-    Pair("XR Uncensored", "xr-uncensored"),
-    Pair("YouMei", "youmei"),
-    Pair("YouMi", "youmi"),
-    Pair("YouMi尤蜜", "youmiapp"),
-    Pair("YouWu", "youwu"),
-    Pair("内购无水印", "内购无水印"),
-    Pair("她们印象", "她们印象"),
-    Pair("精选街拍作品", "精选街拍作品"),
-
-    Pair("--<Korean>--", ""),
-    Pair("AG", "ag"),
-    Pair("Bimilstory", "bimilstory"),
-    Pair("BLUECAKE", "bluecake"),
-    Pair("CreamSoda", "creamsoda"),
-    Pair("DJAWA", "djawa"),
-    Pair("Espacia Korea", "espacia-korea"),
-    Pair("Fantasy Factory", "fantasy-factory"),
-    Pair("Fantasy Story", "fantasy-story"),
-    Pair("Glamarchive", "glamarchive"),
-    Pair("Haivia", "haivia"),
-    Pair("Herovia", "herovia"),
-    Pair("HIGH FANTASY", "high-fantasy"),
-    Pair("KIMLEMON", "kimlemon"),
-    Pair("KIREI", "kirei"),
-    Pair("KiSiA", "kisia"),
-    Pair("Korean Realgraphic", "korean-realgraphic"),
-    Pair("LE", "le"),
-    Pair("Lilynah", "lilynah"),
-    Pair("Lookas", "lookas"),
-    Pair("Loozy", "loozy"),
-    Pair("MakeModel", "makemodel"),
-    Pair("Moon Night Snap", "moon-night-snap"),
-    Pair("Paranhosu", "paranhosu"),
-    Pair("PhotoChips", "photochips"),
-    Pair("Pure Media", "pure-media"),
-    Pair("PUSSYLET", "pussylet"),
-    Pair("SAINT Photolife", "saint-photolife"),
-    Pair("Sera", "sera"),
-    Pair("SWEETBOX", "sweetbox"),
-    Pair("UHHUNG MAGAZINE", "uhhung-magazine"),
-    Pair("UMIZINE", "umizine"),
-    Pair("WXY ENT", "wxy-ent"),
-    Pair("Yo-U", "yo-u"),
-
-    Pair("--<Other>--", ""),
-    Pair("AI Generated", "ai-generated"),
-    Pair("AIGirl", "aigirl"),
-    Pair("Cosplay", "cosplay"),
-    Pair("JP", "jp"),
-    Pair("JVID", "jvid"),
-    Pair("OtherXXX", "otherxxx"),
-    Pair("Patreon", "patreon"),
-    Pair("Private Photoshoot", "private-photoshoot"),
-
-    Pair("--<More>--", ""),
-)
